% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\title{color Class}
\usage{
color(names_for_hex, codes_for_names)
}
\description{
This class provides a way to work with colors in your application.
}
\details{
The color class allows you to get a specific color from the built-in palettes or create a new color palette.
}
\section{Methods}{

  \describe{
    \item{\code{get}}{Retrieve a specific color from the palette.}
  }
}

\section{Properties}{

  \describe{
    \item{\code{names_for_hex}}{A data frame containing color names and their corresponding hex codes.}
    \item{\code{codes_for_names}}{A data frame containing color codes and their corresponding names.}
  }
}

\examples{
co <- RFecharts::color()

co@get(1, alpha =.2)
co@get(1:10)
co@get(c("high", "green"))

# modify the palette
co@names_for_hex <- co@names_for_hex |> 
  dplyr::add_row(color_name = "new_color", hex = "#824bac")

co@get("new_color")

}
