% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_delfi.R
\name{ec_delfi}
\alias{ec_delfi}
\title{ec_delfi}
\usage{
ec_delfi(
  data,
  colors = NULL,
  show_positives = "inclusive",
  nc = RFecharts::name_cleaner(),
  co = RFecharts::color(),
  as_json = shiny::isRunning(),
  theme = RFecharts::rf_echarts_theme
)
}
\arguments{
\item{data}{data.frame containing the data to be visualized.}

\item{colors}{vector of colors to use for the chart.}

\item{show_positives}{string indicating how to show positives. Options are "inclusive", "exclusive", or "none".}

\item{nc}{`name_cleaner` object. Default is `RFecharts::name_cleaner()`.}
}
\value{
echarts4r chart
}
\description{
This function creates an echarts4r chart to display DELFI data.
}
\details{
Visualize DELFI data.
}
\examples{
example_delfi |>
  prep_delfi(ertek_name, count) |> 
  ec_delfi(show_positives = "inclusive")

example_delfi |>
  prep_delfi(ertek_name, count) |> 
  ec_delfi(show_positives = "exclusive")


}
