% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_licert.R
\name{ec_licert}
\alias{ec_licert}
\title{ec_licert}
\usage{
ec_licert(
  data,
  label_type = "ratio",
  xlab = NULL,
  colors = NULL,
  avg_color = "#86868631",
  nc = RFecharts::name_cleaner(),
  co = RFecharts::color(),
  as_json = shiny::isRunning(),
  theme = RFecharts::rf_echarts_theme
)
}
\arguments{
\item{data}{data.frame containing the data to be visualized.}

\item{label_type}{string indicating the type of label to use. Options are "ratio" or "number". Default is "ratio".}

\item{xlab}{string indicating the x-axis label. Default is NULL.}

\item{colors}{vector of colors to use for the chart. Default is NULL.}

\item{avg_color}{string indicating the color to use for the average line. Default is "#86868631".}

\item{nc}{`name_cleaner` object. Default is `RFecharts::name_cleaner()`.}

\item{co}{`color` object. Default is `RFecharts::color()`.}

\item{as_json}{Return as json? By default `shiny::isRunning()`, thus it will appear as chart in your console, but as json in the shiny app.}

\item{theme}{Commonly used RF modification for the end of the chart. Default is `RFecharts::rf_echarts_theme`.}
}
\value{
echarts4r chart
}
\description{
This function creates an echarts4r chart to display license certificate data.
}
\details{
Visualize license certificate data.
}
\examples{
example_licert |>
  prep_licert(group = brand, value = value, n = n) |> 
  ec_licert(xlab = "Rating", avg_color = "red", label_type = "ratio", nc = RFecharts::name_cleaner(language = "eng"))

example_licert |>
  prep_licert(group = brand, value = value, n = n) |> 
  ec_licert(xlab = "Promóciós értékelés", label_type = "count")

}
