% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_means.R
\name{ec_means}
\alias{ec_means}
\title{ec_means}
\usage{
ec_means(
  data,
  y,
  x,
  group,
  colors = NULL,
  xlim = c(0, 5),
  xlab = NULL,
  nc = RFecharts::name_cleaner(),
  co = RFecharts::color(),
  as_json = shiny::isRunning(),
  theme = RFecharts::rf_echarts_theme
)
}
\arguments{
\item{data}{data.frame}

\item{y}{Variable to plot on y axis.}

\item{x}{Variable to plot on x axis.}

\item{group}{Variable to group by (fill).}

\item{colors}{Custom colors. Same length as number of groups.}

\item{xlim}{Default is `c(0, 5)`}

\item{xlab}{Default is `NULL`.}

\item{nc}{`name_cleaner` object. Default is `RFecharts::name_cleaner()`}

\item{co}{`color` object. Default is `RFecharts::color()`}

\item{as_json}{Return as json? By default `shiny::isRunning()`, thus it will appear as chart in your console, but as json in the shiny app.}

\item{theme}{Commonly used RF modification for the end of the chart. Default is `RFecharts::rf_echarts_theme`.}
}
\value{
echarts4r chart
}
\description{
Visualize means of different groups (brands) by several y values (Qs).
}
\examples{
example_means |> 
  ec_means(x = csatorna_egyedi_atlag, group = brand, y = valasz_szovege, xlab = "uj")

}
