% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_prevalent_dist.R
\name{ec_prevalent_dist}
\alias{ec_prevalent_dist}
\title{ec_prevalent_dist}
\usage{
ec_prevalent_dist(
  data,
  colors = NULL,
  nc = RFecharts::name_cleaner(),
  co = RFecharts::color(),
  as_json = shiny::isRunning(),
  theme = RFecharts::rf_echarts_theme
)
}
\arguments{
\item{data}{data.frame containing the data to be visualized.}

\item{colors}{vector of colors to use for the chart. Default is NULL.}

\item{nc}{`name_cleaner` object. Default is `RFecharts::name_cleaner()`.}

\item{co}{`color` object. Default is `RFecharts::color()`.}

\item{as_json}{Return as json? By default `shiny::isRunning()`, thus it will appear as chart in your console, but as json in the shiny app.}

\item{theme}{Commonly used RF modification for the end of the chart. Default is `RFecharts::rf_echarts_theme`.}
}
\value{
echarts4r chart
}
\description{
This function creates an echarts4r chart to display prevalent distribution data.
}
\details{
Visualize prevalent distribution data.
}
\examples{
example_prevalent_dist |>  
  prep_prevalence_dist(IDOSZAK, ADAT_TIPUS, Value) |> 
  ec_prevalent_dist(nc = name_cleaner(language = "hun"))
}
