% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_sov.R
\name{ec_sov}
\alias{ec_sov}
\title{ec_sov}
\usage{
ec_sov(
  data,
  x,
  y,
  group,
  x_label = NULL,
  colors = NULL,
  nc = RFecharts::name_cleaner(),
  co = RFecharts::color(),
  as_json = shiny::isRunning(),
  theme = RFecharts::rf_echarts_theme
)
}
\arguments{
\item{data}{data.frame containing the data to be visualized.}

\item{x_label}{string indicating the x-axis label. Default is NULL.}

\item{colors}{vector of colors to use for the chart. Default is NULL.}

\item{nc}{`name_cleaner` object. Default is `RFecharts::name_cleaner()`.}

\item{co}{`color` object. Default is `RFecharts::color()`.}

\item{as_json}{Return as json? By default `shiny::isRunning()`, thus it will appear as chart in your console, but as json in the shiny app.}

\item{theme}{Commonly used RF modification for the end of the chart. Default is `RFecharts::rf_echarts_theme`.}
}
\value{
echarts4r chart
}
\description{
This function creates an echarts4r chart to display a stacked bar chart with optional percentage labels.
}
\details{
Visualize stacked bar chart with optional percentage labels.
}
\examples{
example_sov |> # ! this ec_ does not have an addtional prep_ function
  ec_sov(x = csatorna_egyedi, y = valasz_szovege, group = brand, nc = name_cleaner(language = "hun"))
}
