% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_formatter.R
\name{formatter}
\alias{formatter}
\title{formatter}
\usage{
formatter(language = "hun", ...)
}
\arguments{
\item{language}{Language. Preferably from `RFecharts::name_cleaner()@language`. Default is "hun".}

\item{...}{Additional parameters passed to `echarts4r::e_axis_formatter()`.}
}
\value{
echarts4r chart
}
\description{
Wrapper for the e_axis_formatter() to suppress warnings and set the language.
}
