% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_concentration.R
\name{prep_concentration}
\alias{prep_concentration}
\title{prep_concentration}
\usage{
prep_concentration(data, n_doctor, n_patient)
}
\arguments{
\item{data}{data.frame containing the data to be prepared.}

\item{n_doctor}{column name or expression for the number of doctors.}

\item{n_patient}{column name or expression for the number of patients.}
}
\value{
data.frame with prepared data
}
\description{
This function prepares the data for visualization by calculating patient per doctor, percent patient, and cumulative patient values.
}
\details{
Prepare data for concentration visualization.
}
\examples{
example_concentration |> 
  prep_concentration(n_doctor = `Orvosok száma`, n_patient = value)

}
