% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_delfi.R
\name{prep_delfi}
\alias{prep_delfi}
\title{prep_delfi}
\usage{
prep_delfi(data, value, n, consensus = -1)
}
\arguments{
\item{data}{data.frame containing the data to be prepared.}

\item{value}{column name or expression for the value to be used for sorting and calculation.}

\item{n}{column name or expression for the count to be used for calculation.}

\item{consensus}{integer indicating the position of the consensus value. Default is -1.}
}
\value{
data.frame with prepared data
}
\description{
This function prepares the DELFI data for visualization by cleaning and formatting the data, calculating percentages, and adding border colors.
}
\details{
Prepare DELFI data for visualization.
}
\examples{
example_delfi |> 
  prep_delfi(value = ertek_name, n = count, consensus = 1)
}
