% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_licert.R
\name{prep_licert}
\alias{prep_licert}
\title{prep_licert}
\usage{
prep_licert(data, group, value, n)
}
\arguments{
\item{data}{data.frame containing the data to be prepared.}

\item{group}{column name or expression for the group variable.}

\item{value}{column name or expression for the value variable.}

\item{n}{column name or expression for the count variable.}
}
\value{
data.frame with prepared data
}
\description{
This function prepares the license certificate data for visualization by selecting the required columns, calculating weighted means, and adding opacity and average labels.
}
\details{
Prepare data for licert visualization.
}
\examples{
example_licert |> 
  prep_licert(group = brand, value = value, n = n)
}
