% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_prevalence_dist.R
\name{prep_prevalence_dist}
\alias{prep_prevalence_dist}
\title{prep_prevalence_dist}
\usage{
prep_prevalence_dist(data, time, type, value)
}
\arguments{
\item{data}{data.frame containing the data to be prepared.}

\item{time}{column name or expression for the time variable.}

\item{type}{column name or expression for the type variable.}

\item{value}{column name or expression for the value variable.}
}
\value{
data.frame with prepared data
}
\description{
This function prepares the prevalent distribution data for visualization by selecting the required columns, filtering out zero values, calculating summary statistics, and binding the results together.
}
\details{
Prepare prevalent distribution data for visualization.
}
\examples{
example_prevalent_dist |>  
  prep_prevalence_dist(IDOSZAK, ADAT_TIPUS, Value)

}
