% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_echarts_theme.R
\name{rf_echarts_theme}
\alias{rf_echarts_theme}
\title{Custom theme for echarts}
\usage{
rf_echarts_theme(chart)
}
\arguments{
\item{chart}{echarts4r chart}
}
\value{
echarts4r chart
}
\description{
Custom theme for echarts
}
